+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
 
 ### Added
 
+- `BincodeSerializer` for `std` feature, which uses [`bincode` crate](https://docs.rs/bincode) for serialization.
+
 - `EncryptedMessage::len()` function to return cipher-text's payload size.
 
 ## [0.3.2] - 2021-06-16