+++ CHANGELOG.md
@@ -8,7 +8,9 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
 
 ### Added
 
-- `BincodeSerializer` for `std` feature, which uses [`bincode` crate](https://docs.rs/bincode) for serialization.
+- New built-in serializers:
+  - `PostcardSerializer`, which uses [`postcard` crate](https://docs.rs/postcard) for serialization.
+  - `BincodeSerializer` for `std` feature, which uses [`bincode` crate](https://docs.rs/bincode) for serialization.
 
 - `EncryptedMessage::len()` function to return cipher-text's payload size.