+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
 
 ## [Unreleased]
 
+## [0.4.0] - 2021-06-17
+
 ### Added
 
 - New built-in serializers:
@@ -56,8 +58,9 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
 [Semantic Versioning]: https://semver.org/
 
 <!-- Versions -->
-[Unreleased]: https://github.com/laysakura/serde-encrypt/compare/0.1.1...HEAD
+[Unreleased]: https://github.com/laysakura/serde-encrypt/compare/0.4.0...HEAD
 [Released]: https://github.com/laysakura/serde-encrypt/releases
+[0.4.0]: https://github.com/laysakura/serde-encrypt/compare/0.3.2...0.4.0
 [0.3.2]: https://github.com/laysakura/serde-encrypt/compare/0.1.1...0.3.2
 [0.2.0]: https://github.com/laysakura/serde-encrypt/compare/0.1.1...0.2.0
 [0.1.1]: https://github.com/laysakura/serde-encrypt/releases/0.1.1